+++ lib/shadcn_ui.dart
@@ -117,6 +117,6 @@ export 'src/utils/border.dart';
 // External libraries
 export 'package:flutter_animate/flutter_animate.dart' hide Effect;
 export 'package:flutter_svg/flutter_svg.dart';
-export 'package:intl/intl.dart';
+export 'package:intl/intl.dart' hide TextDirection;
 export 'package:lucide_icons_flutter/lucide_icons.dart';
 export 'package:two_dimensional_scrollables/two_dimensional_scrollables.dart';