+++ lib/src/components/resizable.dart
@@ -313,9 +313,13 @@ class ShadResizablePanelGroupState extends State<ShadResizablePanelGroup> {
     final effectiveTextDirection =
         widget.textDirection ?? theme.resizableTheme.textDirection;
 
-    final effectivesSizes = panelInfos.length == widget.children.length
+    var effectivesSizes = panelInfos.length == widget.children.length
         ? panelInfos.map((e) => e.size).toList()
         : defaultSizes;
+
+    if (Directionality.of(context) == TextDirection.rtl && isHorizontal) {
+      effectivesSizes = effectivesSizes.reversed.toList();
+    }
 
     final divider = switch (widget.axis) {
       Axis.horizontal => VerticalDivider(
@@ -341,6 +345,7 @@ class ShadResizablePanelGroupState extends State<ShadResizablePanelGroup> {
     return LayoutBuilder(
       builder: (context, constraints) {
         currentConstraints = constraints;
+        final rtl = Directionality.of(context) == TextDirection.rtl;
         Widget child = Flex(
           direction: widget.axis,
           mainAxisAlignment: effectiveMainAxisAlignment,
@@ -388,6 +393,7 @@ class ShadResizablePanelGroupState extends State<ShadResizablePanelGroup> {
           leadingPosition = isHorizontal
               ? leadingPosition * constraints.maxWidth
               : leadingPosition * constraints.maxHeight;
+
           leadingPosition -=
               effectiveDividerSize / 2 + effectiveDividerThickness / 2;
 
@@ -401,7 +407,8 @@ class ShadResizablePanelGroupState extends State<ShadResizablePanelGroup> {
           dividers.add(
             Positioned(
               top: isHorizontal ? 0 : leadingPosition,
-              left: isHorizontal ? leadingPosition : null,
+              left: isHorizontal && !rtl ? leadingPosition : null,
+              right: isHorizontal && rtl ? leadingPosition : null,
               bottom: isHorizontal ? 0 : null,
               child: GestureDetector(
                 onDoubleTap: widget.onDividerDoubleTap ??