+++ lib/src/theme/themes/default_theme_no_secondary_border_variant.dart
@@ -953,7 +953,7 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
       ),
       singleRadius: radius,
       middleRadius: BorderRadius.zero,
-      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
+      padding: const EdgeInsets.symmetric(vertical: 4),
       decoration: ShadDecoration(
         disableSecondaryBorder: true,
         focusedBorder: ShadBorder.all(